move: 설정 관련 파일 Setting 디렉토리로 이동
