--- PawCut/PawCut/Sources/Views/Main/SettingView.swift
@@ -1,59 +1,0 @@
-//
-//  SettingView.swift
-//  PawCut
-//
-//  Created by donghee on 8/19/25.
-//
-
-import SwiftUI
-
-struct SettingView: View {
-    @Environment(\.dismiss) private var dismiss
-    @StateObject private var viewModel = SettingViewModel()
-
-    var body: some View {
-        NavigationStack {
-            ScrollView {
-                VStack(spacing: 8) {
-
-                    Group {
-                        PawSettingRow(title: "프로필 수정") {
-                            viewModel.navigateToProfileEdit()
-                        }
-                        PawSettingRow(title: "소리 수정") {
-                            viewModel.navigateToSoundEdit()
-                        }
-                    }
-                    PawSectionSeparator()
-
-                    Group {
-                        PawSettingRow(title: "1 : 1 문의") {
-                            viewModel.navigateToCustomerInquiry()
-                        }
-                        PawSettingRow(title: "서비스 이용 약관") {
-                            viewModel.navigateToServiceTerms()
-                        }
-                    }
-                }
-                .padding(.top, 8)
-            }
-            .navigationTitle("설정")
-            .navigationBarTitleDisplayMode(.inline)
-            .navigationBarBackButtonHidden(true)
-            .toolbar {
-                ToolbarItem(placement: .navigationBarLeading) {
-                    Button(action: {
-                        dismiss()
-                    }) {
-                        Image(systemName: "chevron.left")
-                            .foregroundColor(.grayScale01)
-                    }
-                }
-            }
-        }
-    }
-}
-
-#Preview {
-    SettingView()
-}
